+++ components/private/Application/src/MaterialApplication.h
@@ -11,4 +11,4 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-#import "UIApplication+AppExtensions.h"
+#import "UIApplication+MDCAppExtensions.h"
