+++ components/private/Application/src/UIApplication+MDCAppExtensions.h
