+++ components/private/Application/src/UIApplication+MDCAppExtensions.m
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#import "UIApplication+AppExtensions.h"
+#import "UIApplication+MDCAppExtensions.h"
 
 @implementation UIApplication (MDCAppExtensions)
